feat: 将FieldSpec类移至独立文件，优化代码结构
